lte:  EPC upgrade
The LTE/EPC model has been enhanced with new features:
    * SGW, PGW and MME are full nodes.
    * There are P2P links between core network nodes.
    * New S5 interface between SGW and PGW nodes based on GTPv2-C protocol.
    * Allow simulations with multiple SGWs and PGWs.

LTE eNB RRC is extended to support:
    * S1 signalling with the core network is initiated after the
      RRC connection establishment procedure is finished.
    * New ATTACH_REQUEST state to wait for finalization of the S1
      signalling with the core network.
    * New InitialContextSetupRequest primitive of the S1 SAP that is
      received by the eNB RRC when the S1 signalling from the core network
      is finished.

Squashed commit of the following from mrequena:epc-upgrade:

    lte: Clarify that e2e is IPv4/IPv6 and core network is IPv4-only
    lte: Remove duplicated example from wscript
    lte: Add new features of the EPC upgrade to CHANGES.html
    lte: Remove duplicated example
    lte: (fixes #3027) S1 signalling is done before RRC connection establishment is finished
    lte: Add new ATTACH_REQUEST state in eNB RRC
    lte: cast a uint8_t when logging
    lte: Fix doxygen comments and doc typos
    lte: Add documentation for EPC split
    lte: Use protocolNumber in new EpcPgwApplication
    lte: Add doxygen doc for EpcSgwApplication class
    lte: Add doxygen doc for EpcPgwApplication class
    lte: Add doxygen doc for EpcMmeApplication class
    lte: PreSerialize/PreDesearialize are only used in the derived messages
    lte: Command line parameters take precedence over parameters in config file
    lte: Rename number of node pairs variable
    lte: Add IPv6 support to new PgwApplication
    lte: Change to new pgw node
    lte: Number of nodes has changed after EPC split
    lte: InitialUeMessage cannot be called if simulation is not started
    lte: Change some names of addresses and clean debug code
    lte: New GetPacketFilters method
    lte: Add new files to build system
    lte: Don't need to build the GtpcHeader by hand, just the message
    lte: GTPC messages derived from GtpcHeader
    lte: Add new PGW Application
    lte: Add new SGW Application
    lte: Add new MME Application
    lte: Add new GTP-C protocol header
    lte: Split SGW/PGW/MME in different nodes
    lte: Add new example

+++ src/lte/model/fdtbfq-ff-mac-scheduler.cc
@@ -636,7 +636,7 @@ FdTbfqFfMacScheduler::DoSchedDlTriggerReq (const struct FfMacSchedSapProvider::S
       newRar.m_grant.m_tpc = 0;
       newRar.m_grant.m_cqiRequest = false;
       newRar.m_grant.m_ulDelay = false;
-      NS_LOG_INFO (this << " UL grant allocated to RNTI " << (*itRach).m_rnti << " rbStart " << rbStart << " rbLen " << rbLen << " MCS " << m_ulGrantMcs << " tbSize " << newRar.m_grant.m_tbSize);
+      NS_LOG_INFO (this << " UL grant allocated to RNTI " << (*itRach).m_rnti << " rbStart " << rbStart << " rbLen " << rbLen << " MCS " << (uint16_t) m_ulGrantMcs << " tbSize " << newRar.m_grant.m_tbSize);
       for (uint16_t i = rbStart; i < rbStart + rbLen; i++)
         {
           m_rachAllocationMap.at (i) = (*itRach).m_rnti;
@@ -1385,7 +1385,7 @@ FdTbfqFfMacScheduler::DoSchedDlTriggerReq (const struct FfMacSchedSapProvider::S
           newDci.m_mcs.push_back (m_amc->GetMcsFromCqi (worstCqi.at (j)));
           int tbSize = (m_amc->GetDlTbSizeFromMcs (newDci.m_mcs.at (j), RgbPerRnti * rbgSize) / 8); // (size of TB in bytes according to table 7.1.7.2.1-1 of 36.213)
           newDci.m_tbsSize.push_back (tbSize);
-          NS_LOG_INFO (this << " Layer " << (uint16_t)j << " MCS selected" << m_amc->GetMcsFromCqi (worstCqi.at (j)));
+          NS_LOG_INFO (this << " Layer " << (uint16_t)j << " MCS selected" << (uint16_t) m_amc->GetMcsFromCqi (worstCqi.at (j)));
           bytesTxed += tbSize;
         }
